Update Time.java
When the time to change is negative, some times the value of seconds and minutes could be negative and it see as xx:xx:0-1,xxx or xx:xx:0-2,xxx; this change pretends to fix that bug. I have assumed that the negative value is less than 60, in other case it requires improve this fix.
Sorry by my english, my native language is spanish.